+++ cpp/tests/io/comp/decomp_test.cu
@@ -15,7 +15,7 @@
  */
 
 #include <io/comp/gpuinflate.h>
-#include <tests/utilities/cudf_test_fixtures.h>
+#include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 #include <vector>