+++ cpp/tests/search/search_test.cu
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-#include <tests/utilities/cudf_test_fixtures.h>
-#include <tests/utilities/column_wrapper.cuh>
-#include <tests/utilities/scalar_wrapper.cuh>
+#include <tests/utilities/legacy/cudf_test_fixtures.h>
+#include <tests/utilities/legacy/column_wrapper.cuh>
+#include <tests/utilities/legacy/scalar_wrapper.cuh>
 #include <cudf/search.hpp>
 
 using cudf::test::column_wrapper;