+++ cpp/tests/types/legacy/dispatcher_test.cu
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include <tests/utilities/cudf_test_fixtures.h>
+#include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
 #include <cudf/cudf.h>